LPS-112426 Renames types.es.js in favor of eventTypes.es.js
