+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/components/PageRenderer/SuccessVariant.es.js
@@ -18,7 +18,7 @@ import ClayLayout from '@clayui/layout';
 import {usePrevious} from 'frontend-js-react-web';
 import React, {useEffect, useState} from 'react';
 
-import {EVENT_TYPES} from '../../actions/types.es';
+import {EVENT_TYPES} from '../../actions/eventTypes.es';
 import {useForm} from '../../hooks/useForm.es';
 import {usePage} from '../../hooks/usePage.es';
 import {setValue} from '../../util/i18n.es';
