+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/components/Pagination.es.js
@@ -16,7 +16,7 @@ import ClayPagination from '@clayui/pagination';
 import classnames from 'classnames';
 import React from 'react';
 
-import {EVENT_TYPES} from '../actions/types.es';
+import {EVENT_TYPES} from '../actions/eventTypes.es';
 import {useEvaluate} from '../hooks/useEvaluate.es';
 import {useForm} from '../hooks/useForm.es';
 import {usePage} from '../hooks/usePage.es';
