+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/hooks/useDrop.es.js
@@ -14,7 +14,7 @@
 
 import {useDrop as useDndDrop} from 'react-dnd';
 
-import {EVENT_TYPES} from '../actions/types.es';
+import {EVENT_TYPES} from '../actions/eventTypes.es';
 import {useForm} from '../hooks/useForm.es';
 import {usePage} from './usePage.es';
