+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/hooks/useForm.es.js
@@ -14,7 +14,7 @@
 
 import React, {useContext, useEffect, useReducer} from 'react';
 
-import {EVENT_TYPES} from '../actions/types.es';
+import {EVENT_TYPES} from '../actions/eventTypes.es';
 import {createReducer} from '../reducers/index.es';
 import {useThunk, useThunkDispatch} from './useThunk.es';
