+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/thunks/fieldChange.es.js
@@ -14,7 +14,7 @@
 
 import {debounce} from 'frontend-js-web';
 
-import {EVENT_TYPES} from '../actions/types.es';
+import {EVENT_TYPES} from '../actions/eventTypes.es';
 import {evaluate} from '../util/evaluation.es';
 import {PagesVisitor} from '../util/visitors.es';
