+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/thunks/nextPage.es.js
@@ -12,7 +12,7 @@
  * details.
  */
 
-import {EVENT_TYPES} from '../actions/types.es';
+import {EVENT_TYPES} from '../actions/eventTypes.es';
 import {evaluate} from '../util/evaluation.es';
 import {PagesVisitor} from '../util/visitors.es';
